requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
dj-database-url==0.5.0
Django==2.2
django-appconf==1.0.4
django-crispy-forms==1.9.0
django-mathfilters==1.0.0
django-registration-redux==2.7
django-tinymce==3.0.2
gunicorn==20.0.4
idna==2.9
Pillow==7.1.2
pinax-referrals==3.0.3
psycopg2==2.8.5
pytz==2019.3
razorpay==1.2.0
requests==2.23.0
rope==0.16.0
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.9
whitenoise==5.0.1
